+++ src/main/java/org/kakaoshare/backend/domain/funding/repository/query/FundingDetailRepositoryCustom.java
@@ -1,6 +1,6 @@
 package org.kakaoshare.backend.domain.funding.repository.query;
 
-import org.kakaoshare.backend.common.vo.Date;
+import org.kakaoshare.backend.common.vo.date.Date;
 import org.kakaoshare.backend.domain.funding.dto.inquiry.ContributedFundingHistoryResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
